sample_gdf: test_kde.py -> conftest.py & reimport
