+++ tests/test_kde.py
@@ -1,20 +1,8 @@
 import shapely
 import pytest
 import geopandas as gpd
-import numpy as np
 from src.kde import calculate_kde_from_gps_points
-
-
-@pytest.fixture
-def sample_gdf():
-    rng = np.random.default_rng(seed=42)  # deterministic
-    x = rng.normal(
-        loc=2_600_000, scale=1000, size=100
-    )  # create sample EPSG:2056 coords
-    y = rng.normal(loc=1_200_000, scale=1000, size=100)
-    geometry = [shapely.Point(xi, yi) for xi, yi in zip(x, y)]
-    return gpd.GeoDataFrame(geometry=geometry, crs="EPSG:2056")
-
+from conftest import sample_gdf
 
 def test_create_static_column_raises_if_data_is_not_geodataframe():
     with pytest.raises(TypeError):
